Code Improvement and Consolidation
